backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic-settings
motor
pymongo
chromadb
openai
google-generativeai
python-dotenv
python-pptx
python-docx
pypdf
httpx
